refactor(utility): WIP separation of concerns module/uow
