+++ utility/unit_of_work/actor.go
@@ -1,4 +1,4 @@
-package utility
+package unit_of_work
 
 // Internal business logic for functionality shared across all `Actors`.
 //
@@ -15,19 +15,19 @@ import (
 
 // Actor setters
 
-func (u *utilityContext) setActorStakeAmount(actorType coreTypes.ActorType, addr []byte, amount *big.Int) typesUtil.Error {
+func (u *baseUtilityUnitOfWork) setActorStakeAmount(actorType coreTypes.ActorType, addr []byte, amount *big.Int) typesUtil.Error {
 	amountStr := utils.BigIntToString(amount)
 
 	var err error
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		err = u.store.SetAppStakeAmount(addr, amountStr)
+		err = u.persistenceRWContext.SetAppStakeAmount(addr, amountStr)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		err = u.store.SetFishermanStakeAmount(addr, amountStr)
+		err = u.persistenceRWContext.SetFishermanStakeAmount(addr, amountStr)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		err = u.store.SetServicerStakeAmount(addr, amountStr)
+		err = u.persistenceRWContext.SetServicerStakeAmount(addr, amountStr)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		err = u.store.SetValidatorStakeAmount(addr, amountStr)
+		err = u.persistenceRWContext.SetValidatorStakeAmount(addr, amountStr)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -38,17 +38,17 @@ func (u *utilityContext) setActorStakeAmount(actorType coreTypes.ActorType, addr
 	return nil
 }
 
-func (u *utilityContext) setActorUnbondingHeight(actorType coreTypes.ActorType, addr []byte, height int64) typesUtil.Error {
+func (u *baseUtilityUnitOfWork) setActorUnbondingHeight(actorType coreTypes.ActorType, addr []byte, height int64) typesUtil.Error {
 	var err error
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		err = u.store.SetAppUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
+		err = u.persistenceRWContext.SetAppUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		err = u.store.SetFishermanUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
+		err = u.persistenceRWContext.SetFishermanUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		err = u.store.SetServicerUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
+		err = u.persistenceRWContext.SetServicerUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		err = u.store.SetValidatorUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
+		err = u.persistenceRWContext.SetValidatorUnstakingHeightAndStatus(addr, height, int32(coreTypes.StakeStatus_Unstaking))
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -59,17 +59,17 @@ func (u *utilityContext) setActorUnbondingHeight(actorType coreTypes.ActorType, 
 	return nil
 }
 
-func (u *utilityContext) setActorPausedHeight(actorType coreTypes.ActorType, addr []byte, height int64) typesUtil.Error {
+func (u *baseUtilityUnitOfWork) setActorPausedHeight(actorType coreTypes.ActorType, addr []byte, height int64) typesUtil.Error {
 	var err error
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		err = u.store.SetAppPauseHeight(addr, height)
+		err = u.persistenceRWContext.SetAppPauseHeight(addr, height)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		err = u.store.SetFishermanPauseHeight(addr, height)
+		err = u.persistenceRWContext.SetFishermanPauseHeight(addr, height)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		err = u.store.SetServicerPauseHeight(addr, height)
+		err = u.persistenceRWContext.SetServicerPauseHeight(addr, height)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		err = u.store.SetValidatorPauseHeight(addr, height)
+		err = u.persistenceRWContext.SetValidatorPauseHeight(addr, height)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -82,19 +82,19 @@ func (u *utilityContext) setActorPausedHeight(actorType coreTypes.ActorType, add
 
 // Actor getters
 
-func (u *utilityContext) getActorStakeAmount(actorType coreTypes.ActorType, addr []byte) (*big.Int, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getActorStakeAmount(actorType coreTypes.ActorType, addr []byte) (*big.Int, typesUtil.Error) {
 	var stakeAmount string
 	var err error
 
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		stakeAmount, err = u.store.GetAppStakeAmount(u.height, addr)
+		stakeAmount, err = u.persistenceReadContext.GetAppStakeAmount(u.height, addr)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		stakeAmount, err = u.store.GetFishermanStakeAmount(u.height, addr)
+		stakeAmount, err = u.persistenceReadContext.GetFishermanStakeAmount(u.height, addr)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		stakeAmount, err = u.store.GetServicerStakeAmount(u.height, addr)
+		stakeAmount, err = u.persistenceReadContext.GetServicerStakeAmount(u.height, addr)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		stakeAmount, err = u.store.GetValidatorStakeAmount(u.height, addr)
+		stakeAmount, err = u.persistenceReadContext.GetValidatorStakeAmount(u.height, addr)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -111,7 +111,7 @@ func (u *utilityContext) getActorStakeAmount(actorType coreTypes.ActorType, addr
 	return amount, nil
 }
 
-func (u *utilityContext) getMaxAllowedPausedBlocks(actorType coreTypes.ActorType) (int, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getMaxAllowedPausedBlocks(actorType coreTypes.ActorType) (int, typesUtil.Error) {
 	var paramName string
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
@@ -126,7 +126,7 @@ func (u *utilityContext) getMaxAllowedPausedBlocks(actorType coreTypes.ActorType
 		return 0, typesUtil.ErrUnknownActorType(actorType.String())
 	}
 
-	maxPausedBlocks, err := u.store.GetIntParam(paramName, u.height)
+	maxPausedBlocks, err := u.persistenceReadContext.GetIntParam(paramName, u.height)
 	if err != nil {
 		return 0, typesUtil.ErrGetParam(paramName, err)
 	}
@@ -134,7 +134,7 @@ func (u *utilityContext) getMaxAllowedPausedBlocks(actorType coreTypes.ActorType
 	return maxPausedBlocks, nil
 }
 
-func (u *utilityContext) getMinRequiredPausedBlocks(actorType coreTypes.ActorType) (int, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getMinRequiredPausedBlocks(actorType coreTypes.ActorType) (int, typesUtil.Error) {
 	var paramName string
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
@@ -149,26 +149,26 @@ func (u *utilityContext) getMinRequiredPausedBlocks(actorType coreTypes.ActorTyp
 		return 0, typesUtil.ErrUnknownActorType(actorType.String())
 	}
 
-	minPausedBlocks, er := u.store.GetIntParam(paramName, u.height)
+	minPausedBlocks, er := u.persistenceReadContext.GetIntParam(paramName, u.height)
 	if er != nil {
 		return 0, typesUtil.ErrGetParam(paramName, er)
 	}
 	return minPausedBlocks, nil
 }
 
-func (u *utilityContext) getPausedHeightIfExists(actorType coreTypes.ActorType, addr []byte) (int64, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getPausedHeightIfExists(actorType coreTypes.ActorType, addr []byte) (int64, typesUtil.Error) {
 	var pauseHeight int64
 	var err error
 
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		pauseHeight, err = u.store.GetAppPauseHeightIfExists(addr, u.height)
+		pauseHeight, err = u.persistenceReadContext.GetAppPauseHeightIfExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		pauseHeight, err = u.store.GetFishermanPauseHeightIfExists(addr, u.height)
+		pauseHeight, err = u.persistenceReadContext.GetFishermanPauseHeightIfExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		pauseHeight, err = u.store.GetServicerPauseHeightIfExists(addr, u.height)
+		pauseHeight, err = u.persistenceReadContext.GetServicerPauseHeightIfExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		pauseHeight, err = u.store.GetValidatorPauseHeightIfExists(addr, u.height)
+		pauseHeight, err = u.persistenceReadContext.GetValidatorPauseHeightIfExists(addr, u.height)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -180,19 +180,19 @@ func (u *utilityContext) getPausedHeightIfExists(actorType coreTypes.ActorType, 
 	return pauseHeight, nil
 }
 
-func (u *utilityContext) getActorStatus(actorType coreTypes.ActorType, addr []byte) (coreTypes.StakeStatus, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getActorStatus(actorType coreTypes.ActorType, addr []byte) (coreTypes.StakeStatus, typesUtil.Error) {
 	var status int32
 	var err error
 
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		status, err = u.store.GetAppStatus(addr, u.height)
+		status, err = u.persistenceReadContext.GetAppStatus(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		status, err = u.store.GetFishermanStatus(addr, u.height)
+		status, err = u.persistenceReadContext.GetFishermanStatus(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		status, err = u.store.GetServicerStatus(addr, u.height)
+		status, err = u.persistenceReadContext.GetServicerStatus(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		status, err = u.store.GetValidatorStatus(addr, u.height)
+		status, err = u.persistenceReadContext.GetValidatorStatus(addr, u.height)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -208,7 +208,7 @@ func (u *utilityContext) getActorStatus(actorType coreTypes.ActorType, addr []by
 	return coreTypes.StakeStatus(status), nil
 }
 
-func (u *utilityContext) getMinRequiredStakeAmount(actorType coreTypes.ActorType) (*big.Int, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getMinRequiredStakeAmount(actorType coreTypes.ActorType) (*big.Int, typesUtil.Error) {
 	var paramName string
 
 	switch actorType {
@@ -224,7 +224,7 @@ func (u *utilityContext) getMinRequiredStakeAmount(actorType coreTypes.ActorType
 		return nil, typesUtil.ErrUnknownActorType(actorType.String())
 	}
 
-	minStake, er := u.store.GetStringParam(paramName, u.height)
+	minStake, er := u.persistenceReadContext.GetStringParam(paramName, u.height)
 	if er != nil {
 		return nil, typesUtil.ErrGetParam(paramName, er)
 	}
@@ -236,7 +236,7 @@ func (u *utilityContext) getMinRequiredStakeAmount(actorType coreTypes.ActorType
 	return amount, nil
 }
 
-func (u *utilityContext) getUnbondingHeight(actorType coreTypes.ActorType) (int64, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getUnbondingHeight(actorType coreTypes.ActorType) (int64, typesUtil.Error) {
 	var paramName string
 
 	switch actorType {
@@ -252,7 +252,7 @@ func (u *utilityContext) getUnbondingHeight(actorType coreTypes.ActorType) (int6
 		return 0, typesUtil.ErrUnknownActorType(actorType.String())
 	}
 
-	unstakingBlocksPeriod, err := u.store.GetIntParam(paramName, u.height)
+	unstakingBlocksPeriod, err := u.persistenceReadContext.GetIntParam(paramName, u.height)
 	if err != nil {
 		return 0, typesUtil.ErrGetParam(paramName, err)
 	}
@@ -260,7 +260,7 @@ func (u *utilityContext) getUnbondingHeight(actorType coreTypes.ActorType) (int6
 	return u.height + int64(unstakingBlocksPeriod), nil
 }
 
-func (u *utilityContext) getMaxAllowedChains(actorType coreTypes.ActorType) (int, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getMaxAllowedChains(actorType coreTypes.ActorType) (int, typesUtil.Error) {
 	var paramName string
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
@@ -273,7 +273,7 @@ func (u *utilityContext) getMaxAllowedChains(actorType coreTypes.ActorType) (int
 		return 0, typesUtil.ErrUnknownActorType(actorType.String())
 	}
 
-	maxChains, err := u.store.GetIntParam(paramName, u.height)
+	maxChains, err := u.persistenceReadContext.GetIntParam(paramName, u.height)
 	if err != nil {
 		return 0, typesUtil.ErrGetParam(paramName, err)
 	}
@@ -281,19 +281,19 @@ func (u *utilityContext) getMaxAllowedChains(actorType coreTypes.ActorType) (int
 	return maxChains, nil
 }
 
-func (u *utilityContext) getActorExists(actorType coreTypes.ActorType, addr []byte) (bool, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getActorExists(actorType coreTypes.ActorType, addr []byte) (bool, typesUtil.Error) {
 	var exists bool
 	var err error
 
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		exists, err = u.store.GetAppExists(addr, u.height)
+		exists, err = u.persistenceReadContext.GetAppExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		exists, err = u.store.GetFishermanExists(addr, u.height)
+		exists, err = u.persistenceReadContext.GetFishermanExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		exists, err = u.store.GetServicerExists(addr, u.height)
+		exists, err = u.persistenceReadContext.GetServicerExists(addr, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		exists, err = u.store.GetValidatorExists(addr, u.height)
+		exists, err = u.persistenceReadContext.GetValidatorExists(addr, u.height)
 	default:
 		return false, typesUtil.ErrUnknownActorType(actorType.String())
 	}
@@ -307,19 +307,19 @@ func (u *utilityContext) getActorExists(actorType coreTypes.ActorType, addr []by
 
 // IMPROVE: Need to re-evaluate the design of `Output Address` to support things like "rev-share"
 // and multiple output addresses.
-func (u *utilityContext) getActorOutputAddress(actorType coreTypes.ActorType, operator []byte) ([]byte, typesUtil.Error) {
+func (u *baseUtilityUnitOfWork) getActorOutputAddress(actorType coreTypes.ActorType, operator []byte) ([]byte, typesUtil.Error) {
 	var outputAddr []byte
 	var err error
 
 	switch actorType {
 	case coreTypes.ActorType_ACTOR_TYPE_APP:
-		outputAddr, err = u.store.GetAppOutputAddress(operator, u.height)
+		outputAddr, err = u.persistenceReadContext.GetAppOutputAddress(operator, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_FISH:
-		outputAddr, err = u.store.GetFishermanOutputAddress(operator, u.height)
+		outputAddr, err = u.persistenceReadContext.GetFishermanOutputAddress(operator, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_SERVICER:
-		outputAddr, err = u.store.GetServicerOutputAddress(operator, u.height)
+		outputAddr, err = u.persistenceReadContext.GetServicerOutputAddress(operator, u.height)
 	case coreTypes.ActorType_ACTOR_TYPE_VAL:
-		outputAddr, err = u.store.GetValidatorOutputAddress(operator, u.height)
+		outputAddr, err = u.persistenceReadContext.GetValidatorOutputAddress(operator, u.height)
 	default:
 		err = typesUtil.ErrUnknownActorType(actorType.String())
 	}
