+++ utility/unit_of_work/application_test.go
@@ -1,4 +1,4 @@
-package utility
+package unit_of_work
 
 import (
 	"testing"
