+++ utility/unit_of_work/message_test.go
@@ -1,4 +1,4 @@
-package utility
+package unit_of_work
 
 import (
 	"encoding/hex"
