+++ utility/unit_of_work/uow_leader.go
@@ -17,6 +17,7 @@ type leaderUtilityUnitOfWork struct {
 func NewForLeader(height int64, readContext modules.PersistenceReadContext, rwPersistenceContext modules.PersistenceRWContext) *leaderUtilityUnitOfWork {
 	return &leaderUtilityUnitOfWork{
 		baseUtilityUnitOfWork: baseUtilityUnitOfWork{
+			height:                 height,
 			persistenceReadContext: readContext,
 			persistenceRWContext:   rwPersistenceContext,
 			logger:                 logger.Global.CreateLoggerForModule(leaderUtilityUOWModuleName),
@@ -24,19 +25,19 @@ func NewForLeader(height int64, readContext modules.PersistenceReadContext, rwPe
 	}
 }
 
-func (uow *leaderUtilityUnitOfWork) CreateProposalBlock(proposer []byte, maxTxBytes uint64, beforeApplyBlock, afterApplyBlock func(modules.UtilityUnitOfWork) error) (stateHash string, txs [][]byte, err error) {
-	if beforeApplyBlock != nil {
-		uow.logger.Debug().Msg("running beforeApplyBlock...")
-		if err := beforeApplyBlock(uow); err != nil {
-			return "", nil, err
-		}
-	}
+func (uow *leaderUtilityUnitOfWork) CreateProposalBlock(proposer []byte, maxTxBytes uint64) (stateHash string, txs [][]byte, err error) {
+	// if burnFn != nil {
+	// 	uow.logger.Debug().Msg("running burnFn...")
+	// 	if err := burnFn(uow); err != nil {
+	// 		return "", nil, err
+	// 	}
+	// }
 
-	if afterApplyBlock != nil {
-		uow.logger.Debug().Msg("running afterApplyBlock...")
-		if err := afterApplyBlock(uow); err != nil {
-			return "", nil, err
-		}
-	}
+	// if rewardFn != nil {
+	// 	uow.logger.Debug().Msg("running rewardFn...")
+	// 	if err := rewardFn(uow); err != nil {
+	// 		return "", nil, err
+	// 	}
+	// }
 	panic("unimplemented")
 }
