+++ utility/unit_of_work/validator.go
@@ -1,4 +1,4 @@
-package utility
+package unit_of_work
 
 // Internal business logic specific to validator behaviour and interactions.
 
@@ -13,7 +13,7 @@ import (
 // This includes validators who double signed, didn't sign at all or disagree with 2/3+ majority.
 // IMPROVE: Need to add more logging to this function.
 // INCOMPLETE: handleByzantineValidators is a WIP and needs to be fully designed, implemented, tested and documented
-func (u *utilityContext) handleByzantineValidators(prevBlockByzantineValidators [][]byte) typesUtil.Error {
+func (u *baseUtilityUnitOfWork) handleByzantineValidators(prevBlockByzantineValidators [][]byte) typesUtil.Error {
 	maxMissedBlocks, err := u.getValidatorMaxMissedBlocks()
 	if err != nil {
 		return err
@@ -21,7 +21,7 @@ func (u *utilityContext) handleByzantineValidators(prevBlockByzantineValidators 
 
 	for _, address := range prevBlockByzantineValidators {
 		// Get the latest number of missed blocks by the validator
-		numMissedBlocks, err := u.store.GetValidatorMissedBlocks(address, u.height)
+		numMissedBlocks, err := u.persistenceReadContext.GetValidatorMissedBlocks(address, u.height)
 		if err != nil {
 			return typesUtil.ErrGetMissedBlocks(err)
 		}
@@ -31,18 +31,18 @@ func (u *utilityContext) handleByzantineValidators(prevBlockByzantineValidators 
 
 		// handle if under the threshold of max missed blocks
 		if numMissedBlocks < maxMissedBlocks {
-			if err := u.store.SetValidatorMissedBlocks(address, numMissedBlocks); err != nil {
+			if err := u.persistenceRWContext.SetValidatorMissedBlocks(address, numMissedBlocks); err != nil {
 				return typesUtil.ErrSetMissedBlocks(err)
 			}
 			continue
 		}
 
 		// pause the validator for exceeding the threshold: numMissedBlocks >= maxMissedBlocks
-		if err := u.store.SetValidatorPauseHeight(address, u.height); err != nil {
+		if err := u.persistenceRWContext.SetValidatorPauseHeight(address, u.height); err != nil {
 			return typesUtil.ErrSetPauseHeight(err)
 		}
 		// update the number of blocks it missed
-		if err := u.store.SetValidatorMissedBlocks(address, numMissedBlocks); err != nil {
+		if err := u.persistenceRWContext.SetValidatorMissedBlocks(address, numMissedBlocks); err != nil {
 			return typesUtil.ErrSetMissedBlocks(err)
 		}
 		// burn validator for missing blocks
@@ -57,7 +57,7 @@ func (u *utilityContext) handleByzantineValidators(prevBlockByzantineValidators 
 // and begins unstaking if the stake falls below the necessary threshold
 // REFACTOR: Extend this to support burning other actors types & pools once the logic is implemented
 // ADDTEST: There are no good tests for this functionality, which MUST be added.
-func (u *utilityContext) burnValidator(addr []byte) typesUtil.Error {
+func (u *baseUtilityUnitOfWork) burnValidator(addr []byte) typesUtil.Error {
 	actorType := coreTypes.ActorType_ACTOR_TYPE_VAL
 	actorPool := coreTypes.Pools_POOLS_VALIDATOR_STAKE
